Issue #3011915: Go to the login form directly to test login errors
It could be that a platform may still want to test whether blocked
users get the correct access denied message but they don't have a
link on the homepage to the login page. This change ensures that
the test also works in that scenario.

Besides making the test more explicit it also removes a step in the
test which speeds up the test.